--- a/test-setup/config/redis.conf
+++ b/test-setup/config/redis.conf
@@ -288,6 +288,7 @@ always-show-logo yes
 #save 900 1
 #save 300 10
 #save 60 10000
+save ""
 
 # By default Redis will stop accepting writes if RDB snapshots are enabled
 # (at least one save point) and the latest background save failed.