Minor change to the resp framing in receive. This also removes an unneeded `Either`. The change is small but according to some profiling it's in a very hot location for the receiving side. Here is a comparison with master (100 wi, 80 i)
```shell
Benchmark                                               Mode  Cnt       Score      Error  Units

Master
LaserdiscCatsByteInByteOutBaselineBench.parallelLoad1  thrpt   80  101429.659 ± 1376.130  ops/s
LaserdiscCatsByteInByteOutBaselineBench.parallelLoad2  thrpt   80  111968.826 ±  775.631  ops/s
LaserdiscCatsByteInRespOutBench.parallelLoad1          thrpt   80   94322.195 ± 1865.012  ops/s
LaserdiscCatsByteInRespOutBench.parallelLoad2          thrpt   80  109039.071 ±  445.255  ops/s


This Branch
LaserdiscCatsByteInByteOutBaselineBench.parallelLoad1  thrpt   80  102043.690 ± 1547.286  ops/s
LaserdiscCatsByteInByteOutBaselineBench.parallelLoad2  thrpt   80  112681.803 ± 1374.423  ops/s
LaserdiscCatsByteInRespOutBench.parallelLoad1          thrpt   80   99965.247 ± 2581.437  ops/s
LaserdiscCatsByteInRespOutBench.parallelLoad2          thrpt   80  119865.775 ± 1071.082  ops/s
```